+++ backend/src/main/java/com/carffeine/carffeine/filter/infrastructure/repository/dto/FilterResponse.java
@@ -1,4 +1,4 @@
-package com.carffeine.carffeine.car.infrastructure.repository.dto;
+package com.carffeine.carffeine.filter.infrastructure.repository.dto;
 
 import com.carffeine.carffeine.filter.domain.Filter;
 import com.carffeine.carffeine.filter.domain.FilterType;
