+++ backend/src/test/java/com/carffeine/carffeine/filter/infrastructure/repository/FilterQueryRepositoryTest.java
@@ -4,11 +4,11 @@ import com.carffeine.carffeine.car.domain.Car;
 import com.carffeine.carffeine.car.domain.CarFilter;
 import com.carffeine.carffeine.car.domain.CarFilterRepository;
 import com.carffeine.carffeine.car.domain.CarRepository;
-import com.carffeine.carffeine.car.infrastructure.repository.dto.FilterResponse;
 import com.carffeine.carffeine.config.QuerydslConfig;
 import com.carffeine.carffeine.filter.domain.Filter;
 import com.carffeine.carffeine.filter.domain.FilterRepository;
 import com.carffeine.carffeine.filter.domain.FilterType;
+import com.carffeine.carffeine.filter.infrastructure.repository.dto.FilterResponse;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
